feat: add sidebar to change background from templates and add favicon
Add Sidebar component to display five preset templates and functionality
to change the background image to the selected template on click.

Update the styles of the layouts of the sidebar and the main content to
use dynamic viewport height instead of screen height for better
responsiveness.

Add global css style and function in the date input to show and remove
the placeholder in phones dynamically.

Add favicon for the app.